+++ checks/datadog.py
@@ -1,4 +1,5 @@
 # stdlib
+import glob
 import os
 import sys
 import traceback
@@ -11,7 +12,10 @@ from itertools import groupby # >= python 2.4
 import modules
 from checks import LaconicFilter
 from checks.utils import TailFile
-from util import windows_friendly_colon_split
+from util import windows_friendly_colon_split, yLoader
+
+# 3rd party
+import yaml
 
 if hasattr('some string', 'partition'):
     def partition(s, sep):
@@ -36,33 +40,10 @@ class Dogstreams(object):
     @classmethod
     def init(cls, logger, config):
         dogstreams_config = config.get('dogstreams', None)
-        dogstreams = []
         if dogstreams_config:
-            # Expecting dogstreams config value to look like:
-            #   <dogstream value>, <dog stream value>, ...
-            # Where <dogstream value> looks like:
-            #   <log path>
-            # or
-            #   <log path>:<module>:<parser function>
-
-            # Create a Dogstream object for each <dogstream value>
-            for config_item in dogstreams_config.split(','):
-                try:
-                    config_item = config_item.strip()
-                    parts = windows_friendly_colon_split(config_item)
-                    if len(parts) == 1:
-                        dogstreams.append(Dogstream.init(logger, log_path=parts[0]))
-                    elif len(parts) == 2:
-                        logger.warn("Invalid dogstream: %s" % ':'.join(parts))
-                    elif len(parts) >= 3:
-                        dogstreams.append(Dogstream.init(
-                            logger,
-                            log_path=parts[0],
-                            parser_spec=':'.join(parts[1:3]),
-                            parser_args=parts[3:],
-                            config=config))
-                except Exception:
-                    logger.exception("Cannot build dogstream")
+            dogstreams = cls._instantiate_dogstreams(logger, config, dogstreams_config)
+        else:
+            dogstreams = []
 
         logger.info("Dogstream parsers: %s" % repr(dogstreams))
 
@@ -71,6 +52,88 @@ class Dogstreams(object):
     def __init__(self, logger, dogstreams):
         self.logger = logger
         self.dogstreams = dogstreams
+
+    @classmethod
+    def _instantiate_dogstreams(cls, logger, config, dogstreams_config):
+        """
+        Expecting dogstreams config value to look like:
+           <dogstream value>, <dog stream value>, ...
+        Where <dogstream value> looks like:
+           <log path>
+        or
+           <log path>:<module>:<parser function>
+        """
+        # Load Dogstream objects from dogstreams.d.
+        dogstreams = cls._load_dogstreams_from_dir(logger, config)
+        # Create a Dogstream object for each <dogstream value>
+        for config_item in dogstreams_config.split(','):
+            try:
+                config_item = config_item.strip()
+                parts = windows_friendly_colon_split(config_item)
+                if len(parts) == 1:
+                    # If the dogstream includes a wildcard, we'll add every
+                    # matching path.
+                    for path in cls._get_dogstream_log_paths(parts[0]):
+                        dogstreams.append(Dogstream.init(logger, log_path=path))
+                elif len(parts) == 2:
+                    logger.warn("Invalid dogstream: %s" % ':'.join(parts))
+                elif len(parts) >= 3:
+                    dogstreams.append(Dogstream.init(
+                        logger,
+                        log_path=parts[0],
+                        parser_spec=':'.join(parts[1:3]),
+                        parser_args=parts[3:],
+                        config=config))
+            except Exception:
+                logger.exception("Cannot build dogstream")
+        return dogstreams
+
+    @classmethod
+    def _load_dogstreams_from_dir(cls, logger, config):
+        dogstreamsd_path = config.get('additional_dogstreamsd', None)
+        if not dogstreamsd_path:
+            return []
+
+        dogstreams = []
+        dogstream_yamls = glob.glob(os.path.join(dogstreamsd_path, "*.yaml"))
+        for dogstream_yaml in dogstream_yamls:
+            parsed = yaml.load(open(dogstream_yaml).read(), Loader=yLoader)
+            dogstreams += cls._dogstream_yaml_to_instance(logger, config, parsed)
+        return dogstreams
+
+    @classmethod
+    def _dogstream_yaml_to_instance(cls, logger, config, parsed):
+        if 'dogstreams' not in parsed:
+            return []
+
+        dogstreams = []
+        for stream in parsed['dogstreams']:
+            # There is only one key in a stream, and it is always its name.
+            stream_name = stream.keys()[0]
+            conf = stream[stream_name]['conf']
+            if 'path' not in conf:
+                logger.error('No path section for dogstream: %s' % stream_name)
+                continue
+            stream_path = conf['path']
+            parser_path = conf.get('parser', None)
+            for log_path in cls._get_dogstream_log_paths(stream_path):
+                dogstream = Dogstream.init(
+                    logger,
+                    log_path=log_path,
+                    parser_spec=parser_path,
+                    parser_args=None,
+                    config=config)
+                dogstreams.append(dogstream)
+        return dogstreams
+
+    @classmethod
+    def _get_dogstream_log_paths(cls, path):
+        """
+        Paths may include wildcard *'s and ?'s.
+        """
+        if '*' not in path:
+            return [path]
+        return glob.glob(path)
 
     def check(self, agentConfig, move_end=True):
         if not self.dogstreams: